# plugins/terraform-backend/src/schema/openapi.yaml
openapi: 3.0.3
info:
  title: Backstage Terraform API
  description: API for making calls to Terraform Enterprise / Cloud from Backstage
  version: 0.0.1
servers:
  - url: /
paths:
  '/health':
    get:
      description: Healthcheck endpoint to ensure the backend is running
      operationId: healthcheck
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/health'
  '/organizations/{orgName}/workspaces/{workspaceNames}/runs':
    get:
      description: Returns data about the recent Terraform Runs
      operationId: getRuns
      parameters:
        - name: orgName
          in: path
          description: Name of organization to return the runs for
          required: true
          schema:
            type: string
        - name: workspaceNames
          in: path
          description: Names of workspace to return the runs for, comma-separated list
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/run'
  '/organizations/{orgName}/workspaces/{workspaceNames}/latestRun':
    get:
      description: Returns data about the most recent Terraform Run
      operationId: getLatestRun
      parameters:
        - name: orgName
          in: path
          description: Name of organization for which to return the latest run
          required: true
          schema:
            type: string
        - name: workspaceNames
          in: path
          description: Names of workspace to return the runs for, comma-separated list
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/latestRun'
  '/organizations/{orgName}/workspaces/{workspaceNames}/assessment-results':
    get:
      description: Returns data about the latest assessment results for one or more workspace
      operationId: getWorkspaceAssessmentResults
      parameters:
        - name: orgName
          in: path
          description: Name of organization
          required: true
          schema:
            type: string
        - name: workspaceNames
          in: path
          description: Names of workspace to return the assessments for, comma-separated list
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/assessmentResults'
  '/terraform-config':
    get:
      description: Returns the Terraform configuration data being used by the plugin.
      operationId: getTerraformConfiguration
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/terraformConfiguration'
components:
  schemas:
    health:
      title: Healthcheck
      description: Healthcheck endpoint
      type: object
      properties:
        status:
          type: string
    run:
      title: Terraform Run
      description: Description of a terraform run
      type: array
      items:
        type: object
        properties:
          id:
            type: string
          message:
            type: string
            nullable: true
          status:
            type: string
          createdAt:
            type: string
          confirmedBy:
            type: object
            nullable: true
            properties:
              user:
                type: string
              avatar:
                type: string
          plan:
            type: object
            nullable: true
            properties:
              logs:
                type: string
          workspace:
            type: object
            nullable: true
            properties:
              name:
                type: string
    latestRun:
      title: Terraform Latest Run
      description: Description of the latest terraform run
      type: object
      properties:
        id:
          type: string
        message:
          type: string
          nullable: true
        status:
          type: string
        createdAt:
          type: string
        confirmedBy:
          type: object
          nullable: true
          properties:
            user:
              type: string
            avatar:
              type: string
        plan:
          type: object
          nullable: true
          properties:
            logs:
              type: string
        workspace:
          type: object
          nullable: true
          properties:
            name:
              type: string
    assessmentResults:
      title: Collection of Terraform Workspace Assessment Results
      type: array
      items:
        $ref: '#/components/schemas/assessmentResult'
    assessmentResult:
      title: Terraform Workspace Assessment Result
      type: object
      properties:
        id:
          type: string
        createdAt:
          type: string
        workspaceId:
          type: string
        workspaceName:
          type: string
        driftMetrics:
          $ref: '#/components/schemas/driftMetrics'
        validationMetrics:
          $ref: '#/components/schemas/validationMetrics'
    driftMetrics:
      title: Workspace Drift Metrics
      type: object
      properties:
        drifted:
          type: boolean
          example: false
        resourcesDrifted:
          type: number
          example: 0
        resourcesUndrifted:
          type: number
          example: 163
    validationMetrics:
      title: Workspace Validation Metrics
      type: object
      properties:
        allChecksSucceeded:
          type: boolean
          example: false
        checksErrored:
          type: number
          example: 0
        checksFailed:
          type: number
          example: 2
        checksPassed:
          type: number
          example: 4
        checksUnknown:
          type: number
          example: 0
    terraformConfiguration:
      title: Terraform Configuration
      type: object
      required:
        - baseUrl
      properties:
        baseUrl:
          type: string
          example: https://app.terraform.io